DRY useAuthenticated
