+++ packages/commerce-sdk-react/src/hooks/useAuthenticatedClient.ts
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2022, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import {
+    QueryFunctionContext,
+    QueryKey,
+    QueryFunction,
+    MutationFunction
+} from '@tanstack/react-query'
+import useAuth from './useAuth'
+import useCommerceApi from './useCommerceApi'
+import {ApiClients, IQueryFunction, IMutationFunction} from './types'
+
+interface Client {
+    clientConfig: {
+        headers: Record<string, string>
+    }
+}
+
+/**
+ * This is a private method. It is a react hook that wraps queries / mutations with
+ * authenticated API clients.
+ *
+ * @internal
+ */
+function useAuthenticatedClient<TData, TVariables = unknown>(
+    fn: IMutationFunction<TData, TVariables>
+): MutationFunction<TData, TVariables>
+function useAuthenticatedClient<TData>(fn: IQueryFunction<TData>): QueryFunction<TData>
+function useAuthenticatedClient<TData, TVariables = unknown>(
+    fn: IQueryFunction<TData> | IMutationFunction<TData, TVariables>
+) {
+    const auth = useAuth()
+    const apiClients = (useCommerceApi() as unknown) as Record<string, Client>
+
+    return (variables: TVariables & QueryFunctionContext<QueryKey>) => {
+        return auth
+            .ready()
+            .then(({access_token}) => {
+                Object.keys(apiClients).forEach((client) => {
+                    apiClients[client].clientConfig.headers = {
+                        ...apiClients[client].clientConfig.headers,
+                        Authorization: `Bearer ${access_token}`
+                    }
+                })
+                return (apiClients as unknown) as ApiClients
+            })
+            .then((apiClients) => fn(variables, apiClients))
+    }
+}
+
+export default useAuthenticatedClient
