+++ packages/commerce-sdk-react/src/hooks/useMutation.ts
@@ -4,13 +4,13 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {useAuthenticatedMutation} from './useAuthenticatedRequest'
-import {useMutation as useReactQueryMutation} from '@tanstack/react-query'
-import {MutationFunction} from './types'
+import useAuthenticatedClient from './useAuthenticatedClient'
+import {useMutation as useReactQueryMutation, MutationKey} from '@tanstack/react-query'
+import {IMutationFunction} from './types'
 
 export const useMutation = <TData = unknown, TError = unknown, TVariables = unknown>(
-    fn: MutationFunction<TData, TVariables>
+    fn: IMutationFunction<TData, TVariables>
 ) => {
-    const authenticatedFn = useAuthenticatedMutation<TData, TVariables>(fn)
+    const authenticatedFn = useAuthenticatedClient<TData, TVariables>(fn)
     return useReactQueryMutation<TData, TError, TVariables>(authenticatedFn)
 }
